requirements.txt
camel-ai>=0.1.0
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
logging>=0.5.1.2
